requirements.txt
cairo-lang
python-dotenv
grpcio
grpcio-tools

